Consolidate and fix forbidden import `pre-commit` checks
